Add support for LEASE frame
Currently, reactivesocket-js doesn't support `LEASE`.
If a client enable the `LEASE` flag during the setup, it will not be
able to parse the `LEASE` frames periodically sent by the server.

Even though nothing is done with the lease, it can still allow a client
to enable it, which can let the server enabling rejection of request when
it is over capacity.