Fix test on OSX.
***Problem***

Currently we're not waiting for the server to receive the connection event
from the client before executing the test.
It's fine on Linux, but OSX is slow enough that the test may run before
the server see the connection event, and then before it initialize the
required global variable.

***Solution***
Explicitely wait that the server and the client are ready before running the
tests. This is done by having a counter, that each side is incrementing, the
first one which reach 2, triggers the `done()` method.