Add full support for lease (#49)
* Add full support for lease

This PR introduce the full support for lease.
Clients can now pass `{lease: true}` when creating a ReactiveSocket,
the client will then wait for the connection to establish but also
wait for the `LEASE` frame.

I added a `availability` method on the `connection`, upon reception
of a `LEASE`, it will update the availability (number between 0 and 1.0).
Note that it doesn't do anything else! It won't prevent the client to send
a message when the budget is empty, but via the `availability` method
it give precious information to a potential higher level library (
e.g. load-balancing library).

There is a simple implementation on the server side, which accept all
clients and give them a budget of 2^30 messages per second.