Introduce timeout on the `Connection`.
***Problem***
Currently there's no notion of timeout, messages sent to a blackhole will never
trigger any events on the client side.

***Solution***
I implemented a new message `CANCEL`, which is send after `requestTimeoutMs`
to the server. The client stream will receive a 'timeout' event.
This change force me to keep the stream object alive a little bit longer,
because now a server can receive another message, even after a `REQUEST`
(e.g. `CANCEL`). The stream is now deleted when either the server respond
or error.

***Modification***
I also moved the `keepalive` and `sendLease` methods from the stream to the connection.
Indeed, those methods must always be used at the connection level (on the stream 0), it
doesn't make sense to let them in the API of stream.