Introduce metrics, a metric event library.
The library is composed of two central objects: the `Recorder` and the `Aggregator`.

You use the `Recorder` to create `Counter`s and `Timer`s, then the `Counter`s and
`Timer`s are used to send events to the `Recorder`.

On the other end, the `Aggregator` listen to events emitted by the `Counter`s / `Timer`s
and aggregate them. The default `Aggregator` aggregates counter events as a single
value (the latest) and timer event into an histogram.
But nothing is preventing you for having different aggregating strategy, for
instance you may want to keep a min/avg/latest/max structure for every counters.

The library provides two histograms implementations, one based on the classical
bucketed algorithm and one based on the more inovative frugal streaming
estimation. The first one is as precise as you want with an impact on speed and
memory footprint, the second one is extremly fast with almost no memory impact
but is less precise.