Refactor frame parsing/encoding to avoid copying buffers.
The current implementation create network by concatenating a series of small
buffer.
Under the hood, Buffer.concat is actually allocating a new bigger buffer and
copying the data from old to the new buffer.

This can be avoided by computing up front the size of the buffer, and writing
directly to it. This saves allocations and reduce GC pressure as well as
improving the overall throughput.

This refactoring make the rsBench/tcpServer bench ~40% faster.