ReactiveSocket deal with stream error
Previously when the underlying transport stream was emitting an error, the system was crashing since nobody was listening to the event.

Now the ReactiveSocket class listen to the event and add extra protection for updating the availability right away.
The loadbalancer remove the ReactiveSocket from its list as soon as it errors.