Fix bug in utf8 data length computation
The current version of the code use `string.length` to compute the length of
the data/metadata. This is obviously wrong in the case of utf8 strings, we need
to use `Buffer.byteLength(string, encoding)` instead.

I also added two test that ensure that data/metadata are correctly
serialize/deserialize.