Introduce ReEnqueueFilter
This filter is responsible for re-enqueing (retrying) requests that are safe to
re-enqueue (e.g. REJECTED requests or requestsnot transmitted to the server
because of transport error).

The filter also compute a moving average of the re-enqueue rate and stop
re-enqueuing if this rate crosses a treshold (default 5%) to avoid overwhelming
the server with re-enqueue requests (problem also infamously known as retry
storm).